Added class and init to help command
