+++ run.py
@@ -52,6 +52,7 @@ async def load_extensions():
     extensions = [
         "BirthdayBot.Cogs.Registration",
         "BirthdayBot.Cogs.UserAgeInfo",
+        "BirthdayBot.Cogs.Help",
     ]
     for filename in extensions:
         await bot.load_extension(filename)
